// src/lib/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
// ...
      projects: {
        Row: {
          created_at: string | null;
          description: string | null;
          id: string;
          is_favorite: boolean | null;
          name: string;
          profit: number | null;
          quantity: number | null;
          sale_price: number | null;
          selling_price: number | null;
          total_cost: number | null;
          user_id: string | null;
        };
        Insert: {
          created_at?: string | null;
          description?: string | null;
          id?: string;
          is_favorite?: boolean | null;
          name: string;
          profit?: number | null;
          quantity?: number | null;
          sale_price?: number | null;
          selling_price?: number | null;
          total_cost?: number | null;
          user_id?: string | null;
        };
        Update: {
          created_at?: string | null;
          description?: string | null;
          id?: string;
          is_favorite?: boolean | null;
          name?: string;
          profit?: number | null;
          quantity?: number | null;
          sale_price?: number | null;
          selling_price?: number | null;
          total_cost?: number | null;
          user_id?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};